+++ tests/FeatureTests/Security/Scenario07SearchAccessTests/_07_01_SearchAccessTest.php
@@ -33,7 +33,7 @@ class _07_01_SearchAccessTest extends BaseRequestTestCase
             ]
         );
 
-        $this->assertIsUnifiedCollectionResponse(
+        $this->assertIsElementCollectionResponse(
             $response,
             3,
             [